add some static_casts to silence warnings
MSVC is paranoid of casts, indicating possible data truncation due to the
conversion.  Add explicit static casts and assertions as necessary to ensure
that we catch any truncation where it matters.

+++ Sources/Log.cpp
@@ -34,7 +34,7 @@ void vLog(int category, int level, char const *classname, char const *funcname,
   if (sLogLevel > level)
     return;
 
-  if ((sLogMask & (1 << category)) == 0)
+  if ((sLogMask & static_cast<uint64_t>(1 << category)) == 0)
     return;
 
   std::stringstream ss;
